app directory structure
Within the MDK we would like to serve html, css, and images from /app/, but it gets a bit gnarly because we don't want to expose the entire project directory. If we were to move all of the relevant files into an ./app directory, it would be much easier to separate unbuilt files from built files from miscellaneous stuff. Proposed structure would look like this: