ignore app/img file changes when reloading models
Unfortunately there's a common use case where a user edits the Cold State, then goes to Media Picker, and uploads an image. The act of uploading the image triggers a 'something changed' websocket action, which reloads the model and discards the user's changes.

For now, we'll ignore app/img changes until we can actually get back which files changed from the LiveReload server.